+++ apps/api-journeys/src/app/modules/action/action.module.ts
@@ -3,11 +3,11 @@ import { DatabaseModule } from '@core/nest/database'
 import { DateTimeScalar } from '../../lib/dateTime/dateTime.provider'
 import { JourneyService } from '../journey/journey.service'
 import { BlockService } from '../block/block.service'
-import { ActionResolver } from './action.resolvers'
-import { NavigateToJourneyActionResolver } from './navigateToJourney/navigateToJourney.resolvers'
-import { NavigateToBlockActionResolver } from './navigateToBlockAction/navigateToBlockAction.resolvers'
-import { LinkToActionResolver } from './linkToAction/linkToAction.resolvers'
-import { NavigateActionResolver } from './navigateAction/navigateAction.resolvers'
+import { ActionResolver } from './action.resolver'
+import { NavigateToJourneyActionResolver } from './navigateToJourney/navigateToJourney.resolver'
+import { NavigateToBlockActionResolver } from './navigateToBlockAction/navigateToBlockAction.resolver'
+import { LinkToActionResolver } from './linkToAction/linkToAction.resolver'
+import { NavigateActionResolver } from './navigateAction/navigateAction.resolver'
 
 @Module({
   imports: [DatabaseModule],
